Focused the Document Item class and implemented early JSON serialization
Modified:
Moved the Key property from [DocumentItem] to the derived class [Section] as it wouldn't be necessary to have all DocumentItems bound to a key since Sections are what will most likely be file referenced.

Added:
Started the unit test project for this overall project in order to perform iterative testing and rolling validation.