Implemented constructor tests on Section class
Broke Key out of the SectionItem since groups of sections won't hold section objects inside them but rather they'll hold section references based on the key.

Identified the cause of the error in the last commit which was due to the structure of the Key property.  Key properties cannot be self-referencing.

Implemented tests in SectionTests that test the constructors for the Section class

Will need to re-assess the implementation of deserialization to allow for error handling.